Script.
IT Creative Coffee project logic illustration
Written in Grade 10
Cieas Chen


一：branch logic：
From these five variables containing (" yes "or" no "is two branches), the framework of the whole program is constructed, which is initially to ask whether to drink coffee or not. In order to keep the train of thought relatively orderly and convenient for later modification, two branches of y=yes and no were started to be separated, and then two branches of z could be separated from the two branches, and a could be divided from the branches of z. If this was repeated, what would be found missing. Yes, this only ensures the completion of the feeder of each no. We also need to consider the completion of the feeder of the total yes and the feeder of yes in no. So the whole idea of the program is the total score, the final total of 64 branches, 64 output. Details of feeder train of thought:
If the format is a mess, if the branch line write less, check up must be difficult and laborious......
Ii. How to output corresponding documents:
I put nearly 100 kinds of coffee in TXT file. ), each of which has its own pair code.
For example to drink Blackcoffee: his secret code (1 20 100 1 2 Blackcoffee)
Input yes to the question of whether to add water, and after code processing, the variable corresponding to water assigned by the system is 1. 

As shown in the figure below:
Enter yes, and the corresponding variable will not be assigned temporarily. After waiting for the user to input milk amount XXml, input and eval will be assigned to the variable. (PS: if no is entered into this problem, the system will match the second item in the text which is 2) as shown below:
100: milk foam process is the same as adding milk (PS is the same).
1. Caramel and water are the same.
2: in the same way

Three: increase the winning rate setting:

In order to prevent coffee cows from crashing coffee varieties in the case of plus or minus 10, the idea of writing text is to construct a difference of 20ml in the amount of milk of each kind.

Turtle photo:
Create a function to draw a cup, divided into large, small and medium, the principle is very simple, in the 64 results of the export of the application function let the user input corresponding to the large and small middle string can (eval, input corresponding to the function starting variable), pigment selection, use fillcolor input the same.

V. final operation screenshot: